ad_email_aliases/requirements.txt
# Python libraries
python-ldap
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
